Renaming module profile to profile-impl
