+++ features/profile-impl/src/main/AndroidManifest.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="com.orafaraujo.features.profileimpl">
+
+    <application>
+        <activity android:name="com.orafaraujo.features.profileimpl.ProfileActivity" />
+    </application>
+
+</manifest>
