+++ features/profile-impl/src/main/java/com/orafaraujo/features/profileimpl/ProfileMapper.kt
@@ -1,0 +1,13 @@
+package com.orafaraujo.features.profileimpl
+
+import com.orafaraujo.usecases.loadprofileapi.models.ProfileDTO
+
+class ProfileMapper {
+
+    fun toPresentationModel(profile: ProfileDTO): ProfilePresentationModel {
+        return ProfilePresentationModel(
+            id = profile.id,
+            name = profile.name
+        )
+    }
+}
