+++ features/profile-impl/src/main/java/com/orafaraujo/features/profileimpl/ProfilePresentationModel.kt
@@ -1,0 +1,6 @@
+package com.orafaraujo.features.profileimpl
+
+data class ProfilePresentationModel(
+    val id: Long,
+    val name: String,
+)
