--- features/profile/src/main/AndroidManifest.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.orafaraujo.features.profile">
-
-    <application>
-        <activity android:name=".ProfileActivity" />
-    </application>
-
-</manifest>
