--- features/profile/src/main/java/com/orafaraujo/features/profile/ProfileMapper.kt
@@ -1,13 +1,0 @@
-package com.orafaraujo.features.profile
-
-import com.orafaraujo.usecases.loadprofileapi.models.ProfileDTO
-
-class ProfileMapper {
-
-    fun toPresentationModel(profile: ProfileDTO): ProfilePresentationModel {
-        return ProfilePresentationModel(
-            id = profile.id,
-            name = profile.name
-        )
-    }
-}
