--- features/profile/src/main/java/com/orafaraujo/features/profile/ProfilePresentationModel.kt
@@ -1,6 +1,0 @@
-package com.orafaraujo.features.profile
-
-data class ProfilePresentationModel(
-    val id: Long,
-    val name: String,
-)
